Se continuaron con validaciones

+++ src/Components/BotonEliminarUsuario.jsx
@@ -1,11 +1,26 @@
 import axios from 'axios'
 import Swal from 'sweetalert2'
 import { useNavigate } from 'react-router-dom'
+import { useState, useEffect } from 'react'
 
-export const BotonEliminarUsuario = ({idUsuario, nombre}) => {
+export const BotonEliminarUsuario = ({idUsuario, nombre, usuario}) => {
   const navigate = useNavigate()
+  const token = localStorage.getItem('token')
+  const [usuarioEsp, setUsuarioEsp] = useState({})
+  const idUsuarioLogueado = localStorage.getItem("idUsuario");
 
-  const borrarUsuario = () => {
+  useEffect(() => {
+    if(idUsuarioLogueado){
+      const obtenerUsuario = async () => {
+        const respuesta = await axios.get(`http://localhost:8000/usuarios/${idUsuarioLogueado}`);
+        setUsuarioEsp(respuesta.data.usuario);
+      }
+
+      obtenerUsuario()
+    }
+  }, [])
+
+  const borrarUsuario = (nombre) => {
     Swal.fire({
       title: `¿Estás seguro de que quieres eliminar a ${nombre}?`,
       text: "No se puede revertir el proceso en caso de eliminar el usuario",
@@ -16,24 +31,42 @@ export const BotonEliminarUsuario = ({idUsuario, nombre}) => {
       confirmButtonText: "Si, borrar!"
     }).then(async (result) => {
       if(result.isConfirmed){
-        const respuesta = await axios.delete("http://localhost:8000/usuarios/eliminar-usuario", {
-          data: {
-            id: idUsuario
-          }
-        })
-        console.log(respuesta)
-
-        if(respuesta.data.status === 200){
+        if(usuario === "admin.ecomrc"){
           Swal.fire({
-            icon:'success',
-            title: respuesta.data.message,
-            showConfirmButton: false,
-            timer: 1500
+            title: "No se puede eliminar al administrador general de la página",
+            icon: "warning",
+            showConfirmButton: true
           })
-
-          setTimeout(() => {
-            navigate(0)
-          }, 1500);
+        }
+        else{
+          const respuesta = await axios.delete("http://localhost:8000/usuarios/eliminar-usuario", {
+            data: {
+              id: idUsuario,
+              accessToken: token
+            }
+          })
+          console.log(respuesta)
+  
+          if(respuesta.data.status === 200){
+            Swal.fire({
+              icon:'success',
+              title: respuesta.data.message,
+              showConfirmButton: false,
+              timer: 1500
+            })
+  
+            setTimeout(() => {
+              navigate(0)
+            }, 1500);
+          }
+          else if(respuesta.data.status === 500){
+            Swal.fire({
+              icon: 'error',
+              title: "La accion no se puede realizar porque el token expiró o es inexistente",
+              showConfirmButton: false,
+              timer: 1500
+            })
+          }
         } 
       }
     })
@@ -41,9 +74,19 @@ export const BotonEliminarUsuario = ({idUsuario, nombre}) => {
 
   return (
     <>
-      <button className="btn btn-danger btn-sm" onClick={() => borrarUsuario(idUsuario, nombre)}> 
-        <i className="bi bi-trash2-fill"></i>
-      </button>
+      {
+        usuarioEsp.estado === "Pendiente" ?
+        <button className="btn btn-danger btn-sm" onClick={() => borrarUsuario(nombre)} disabled> 
+          <i className="bi bi-trash2-fill"></i>
+        </button>
+        :
+        usuarioEsp.estado === "Activo" ?
+        <button className="btn btn-danger btn-sm" onClick={() => borrarUsuario(nombre)}> 
+          <i className="bi bi-trash2-fill"></i>
+        </button>
+        :
+        null
+      }
     </>
   )
 }
